Fix build errors when building for iOS
Since commit 62b18db041abd38c728e124cd952fa6c08363b73, builds fail when
switching between building for the iOS simulator and building for iOS
devices, and vice versa.

Prior to that commit, the linker would write libarc.a to an
architecture-dependent path
(Build/Intermediates/arc.build/Debug-iphonesimulator/ for the simulator,
Build/Intermediates/arc.build/Debug-iphoneos/ for devices). It now
writes to a single, unified path. That means that if you've previously
built for a device (armv7/armv7s), the linker will try to use that
binary when you then build for the simulator (i386), and vice versa.

I don't know what the motivation behind the original commit was, but I'd
suggest it's rolled back pending a rethink.